requirements.txt
alembic==1.18.4
blinker==1.9.0
click==8.3.1
croniter==6.0.0
factory_boy==3.3.3
Faker==40.4.0
Flask==3.1.3
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.2.0
rq==2.7.0
six==1.17.0
SQLAlchemy==2.0.46
typing_extensions==4.15.0
Werkzeug==3.1.6
